+++ VAX/vax630_io.c
@@ -557,7 +557,7 @@ if ((ba | bc) & 03) {                                   /* check alignment */
             if (!qba_map_addr (ba + i, &ma))            /* inv or NXM? */
                 return (bc - i);
             }
-        *buf = ReadB (ma);
+        *buf = (uint8)ReadB (ma);
         ma = ma + 1;
         }
     }
@@ -591,7 +591,7 @@ if ((ba | bc) & 03) {                                   /* check alignment */
             if (!qba_map_addr (ba + i, &ma))            /* inv or NXM? */
                 return (bc - i);
             }
-        *buf = ReadW (ma);
+        *buf = (uint16)ReadW (ma);
         ma = ma + 2;
         }
     }
